feat: move CustomExtraField to own file
